refactor(ui): 统一并语义化UI Handler接口
主要变更：

1.  **统一接口**: 在 `UIHandlerBase` 中定义了一套全新的、基于语义的抽象方法（如 `print_success`, `print_banner`, `get_input_semantic` 等）。
2.  **实现新接口**: `RichHandler` 和 `FletHandler` 都完整实现了这套新的语义化接口，使得不同UI后端的行为高度一致。
3.  **代码整合**: 删除了 `rich_handler_new.py`，将其内容完全整合进 `rich_handler.py`，并重构其为单例模式，统一了实现。
4.  **保留兼容性**: 所有旧的UI方法（如 `print_text`, `get_input`）在各个实现中仍然保留，确保了对现有代码的向后兼容性。
5.  **更新调用方**: `PageBase` 等核心模块现在统一调用新的语义化接口，代码更清晰，意图更明确。

此次重构极大地提升了UI层的抽象水平和代码一致性，为未来维护和扩展新的UI后端（如Web UI）奠定了坚实的基础。
